+++ package.json
@@ -43,8 +43,7 @@
     "slick-carousel": "^1.8.1",
     "typescript": "^4.9.5",
     "validator": "^13.12.0",
-    "web-vitals": "^2.1.4",
-    "xlsx": "^0.18.5"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
