+++ src/Layout/Venue/Space/SpaceList/space.scss
@@ -1,5 +1,5 @@
-@import "../../../styles/mixins";
-@import "../../../styles/color.scss";
+@import "../../../../styles/mixins";
+@import "../../../../styles/color.scss";
 
 .space {
   &-row {
