+++ src/Layout/Venue/Venue/index.tsx
@@ -1,0 +1,7 @@
+export const VenueOption = () => {
+  return (
+    <div>
+      <h1>Venue Option</h1>
+    </div>
+  );
+};
