+++ src/Layout/Venue/index.tsx
@@ -1,46 +1,46 @@
+import { Menu } from "antd";
+import { MenuProps } from "antd/lib";
+import React from "react";
 import { FooterWeb } from "../../LayoutOption/FooterWeb";
 import { HeaderNavBar } from "../../LayoutOption/HeaderNavBar";
 import "./style.scss";
-import ScrollSpyLayout from "../../LayoutOption/ScrollSpyLayout";
-import { Space } from "./Space";
-export const Venue = () => {
-  const menuItems = [
+
+type MenuItem = Required<MenuProps>["items"][number];
+const Venue: React.FC = () => {
+  const items: MenuItem[] = [
     {
       key: "1",
       icon: (
         <img src="https://img.icons8.com/windows/32/company.png" alt="space" />
       ),
-      label: "Không gian",
-      children: [
-        { key: "1-1", label: "Thông tin chi tiết" },
-        { key: "1-2", label: "Thông tin khách" },
-      ],
+      label: "Tổng quan",
     },
     {
       key: "2",
       icon: (
         <img src="https://img.icons8.com/small/32/company.png" alt="venue" />
       ),
-      label: "Địa điểm",
-      isParent: true,
-      children: [
-        { key: "2-1", label: "Thông tin chi tiết" },
-        { key: "2-2", label: "Thông tin khách" },
-      ],
+      label: "Giới thiệu",
     },
   ];
-
-  const contentSections = {
-    "1-1": <Space />,
-    "1-2": <Space />,
-    "2-1": <Space />,
-    "2-2": <Space />,
+  const onClick: MenuProps["onClick"] = (e) => {
+    console.log("click ", e);
   };
+
   return (
-    <div className="venue">
+    <div className="venue-layout">
       <HeaderNavBar isLogin={true} />
-      <ScrollSpyLayout items={menuItems} contentSections={contentSections} />
+      <Menu
+        onClick={onClick}
+        style={{ width: 256 }}
+        defaultSelectedKeys={["1"]}
+        defaultOpenKeys={["1"]}
+        mode="inline"
+        items={items}
+      />
       <FooterWeb />
     </div>
   );
 };
+
+export default Venue;
